# docs/oracle-api/swagger.yaml
basePath: /api
definitions:
  BridgingRequestStateResponse:
    properties:
      destinationChainId:
        description: Destination chain ID
        type: string
      destinationTxHash:
        description: Destination transaction hash
        type: string
      sourceChainId:
        description: Source chain ID
        type: string
      sourceTxHash:
        description: Source transaction hash
        type: string
      status:
        allOf:
        - $ref: '#/definitions/BridgingRequestStatus'
        description: Status of bridging request
    type: object
  BridgingRequestStatus:
    enum:
    - DiscoveredOnSource
    - InvalidRequest
    - SubmittedToBridge
    - IncludedInBatch
    - SubmittedToDestination
    - FailedToExecuteOnDestination
    - ExecutedOnDestination
    type: string
    x-enum-varnames:
    - BridgingRequestStatusDiscoveredOnSource
    - BridgingRequestStatusInvalidRequest
    - BridgingRequestStatusSubmittedToBridge
    - BridgingRequestStatusIncludedInBatch
    - BridgingRequestStatusSubmittedToDestination
    - BridgingRequestStatusFailedToExecuteOnDestination
    - BridgingRequestStatusExecutedOnDestination
  CardanoChainConfigUtxo:
    properties:
      address:
        description: Output address
        type: string
      amount:
        description: Amount of currency
        type: integer
      id:
        description: Transaction hash
        items:
          type: integer
        type: array
      index:
        description: Output index
        type: integer
      slot:
        description: Output slot
        type: integer
      tokens:
        description: Tokens
        items:
          $ref: '#/definitions/wallet.TokenAmount'
        type: array
    type: object
  ErrorResponse:
    properties:
      err:
        type: string
    type: object
  HasTxFailedResponse:
    properties:
      failed:
        description: true if the transaction failed, false otherwise
        type: boolean
    type: object
  OracleStateResponse:
    properties:
      chainID:
        description: Chain ID
        type: string
      hash:
        description: Latest block hash
        type: string
      slot:
        description: Latest block slot
        type: integer
      utxos:
        description: Unspent transaction outputs
        items:
          $ref: '#/definitions/CardanoChainConfigUtxo'
        type: array
    type: object
  SettingsResponse:
    properties:
      maxAmountAllowedToBridge:
        description: Maximum amount of currency allowed to be bridged
        type: string
      maxReceiversPerBridgingRequest:
        description: Maximum number of receivers allowed in a bridging request
        type: integer
      maxTokenAmountAllowedToBridge:
        description: Maximum amount of native tokens allowed to be bridged
        type: string
      minChainFeeForBridging:
        additionalProperties:
          type: integer
        description: For each chain, the minimum fee required to cover the submission
          of the transaction on the destination chain
        type: object
      minOperationFee:
        additionalProperties:
          type: integer
        description: For each chain, the minimum fee required to cover operational
          costs
        type: object
      minUtxoChainValue:
        additionalProperties:
          type: integer
        description: For each chain, the minimum allowed UTXO value
        type: object
      minValueToBridge:
        description: Minimum value allowed to be bridged
        type: integer
    type: object
  wallet.TokenAmount:
    properties:
      nam:
        description: name must plain name and not be hex encoded
        type: string
      pid:
        type: string
      val:
        type: integer
    type: object
info:
  contact: {}
  title: Oracle API
  version: "1.0"
paths:
  /BridgingRequestState/Get:
    get:
      description: Returns the current status of a bridging request, along with the
        destination chain ID and transaction hash, based on the given source chain
        ID and transaction hash.
      parameters:
      - description: Source chain ID
        in: query
        name: chainId
        required: true
        type: string
      - description: Source transaction hash
        in: query
        name: txHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Returns the bridging request with its current state.
          schema:
            $ref: '#/definitions/BridgingRequestStateResponse'
        "400":
          description: Bad Request – One or more query parameters are missing, or
            the bridging request state could not be retrieved.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized – API key missing or invalid.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found - No bridging request found for the given parameters.
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get state of bridging request
      tags:
      - BridgingRequestState
  /BridgingRequestState/GetMultiple:
    get:
      description: Returns statuses and related data for one or more bridging requests,
        based on the given source chain ID and transaction hash(es).
      parameters:
      - description: Source chain ID
        in: query
        name: chainId
        required: true
        type: string
      - collectionFormat: multi
        description: Source transaction hashes
        in: query
        items:
          type: string
        name: txHash
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK – Returns a map with source transaction hashes as keys and
            the associated bridging request data as values.
          schema:
            additionalProperties:
              $ref: '#/definitions/BridgingRequestStateResponse'
            type: object
        "400":
          description: Bad Request – chainId is missing from the query or the bridging
            request states could not be retrieved.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized – API key missing or invalid.
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get states of multiple bridging requests
      tags:
      - BridgingRequestState
  /OracleState/Get:
    get:
      description: Returns all unspent transaction outputs of bridging addresses for
        the given chain ID. Each UTXO includes the address it belongs to, along with
        amount and other relevant data.
      parameters:
      - description: Chain ID
        in: query
        name: chainId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Returns Oracle state.
          schema:
            $ref: '#/definitions/OracleStateResponse'
        "400":
          description: Bad Request – chainId is missing from the query or invalid,
            or the transaction outputs could not be retrieved.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized – API key missing or invalid.
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get unspent transaction outputs of bridging addresses on specific chain
      tags:
      - OracleState
  /OracleState/GetHasTxFailed:
    get:
      description: Returns whether the transaction failed. A transaction is considered
        failed if it is not found on Oracle and if its TTL has expired.
      parameters:
      - description: Source chain ID
        in: query
        name: chainId
        required: true
        type: string
      - description: Source transaction hash
        in: query
        name: txHash
        required: true
        type: string
      - description: Transaction TTL
        in: query
        name: ttl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Returns transaction failure status.
          schema:
            $ref: '#/definitions/HasTxFailedResponse'
        "400":
          description: Bad Request – One or more query parameters are missing or invalid,
            or the transaction could not be retrieved.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized – API key missing or invalid.
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get transaction failure status
      tags:
      - OracleState
  /Settings/Get:
    get:
      description: Returns configuration settings for the bridge, including global
        parameters such as minimum and maximum bridging amounts, as well as any per-chain
        specific settings.
      produces:
      - application/json
      responses:
        "200":
          description: OK - Returns bridge configuration.
          schema:
            $ref: '#/definitions/SettingsResponse'
        "401":
          description: Unauthorized – API key missing or invalid.
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get bridge settings for all supported chains
      tags:
      - Settings
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"